fix DT 1701

+++ database/migrations/2023_00_00_000000_create_items_table.php
@@ -68,7 +68,7 @@ class CreateItemsTable extends Migration
 			//
 			$table->unsignedInteger('tag_id')->nullable();
 			
-			$table->unsignedInteger('category_id')->nullable();
+			$table->unsignedBigInteger('category_id')->nullable();
             $table->index('category_id', 'item_category_idx');
             $table->foreign('category_id', 'item_category_fk')->on('categories')->references('id');
